fixed issue in productcontroller file

+++ backend/src/main/java/com/kcare/kcare/Product/Model/Product.java
@@ -31,17 +31,16 @@ public class Product extends BaseEntity {
     private String productName;
     private String productDescription;
     private String productCategory; // asset/cosumable
-    private Integer buyingPrice; // ! sgst + cgst or igst
+    private Double buyingPrice; // ! sgst + cgst or igst
     private Integer quantity;
     private LocalDate expiryDate;
     private Integer thresholdValue;
     private boolean isContainSubpart;
     private Integer sgstTaxPercent;
     private Double totalsgstTaxtAmount;
-    private Double cgstTaxPercent;
+    private Integer cgstTaxPercent;
     private Double totalcgstTaxAmount;
-
-    private Double igstTaxPercent;
+    private Integer igstTaxPercent;
     private Double totaligstTaxAmount;
     private Double taxableAmount; // ! taxableAmount
